requirements.txt
Flask
flasgger
gunicorn
openai-whisper
torch
transformers
pytest
pytest-cov

